Use a single variable object for each name during parsing
This ensures that multiple occurrences of the same variable get
represented by the same object, which is required to make unification
work right. That’s because unification uses variable identity rather than
name, and _that’s_ because variables need to be scoped to the rule in
which they appear — two rules with a variable called t₁ shouldn’t affect
each other when they appear within the same derivation.

(We’re not yet doing anything about multiple instances of the same rule,
but there aren’t any tests which require that yet.)