Add syntactic restrictions through additional inference rules
We should also be checking that each t₁, t₂, … is a syntactically valid
term; this isn’t useful while the parser is producing the formulae, but
if we move to raw arrays of symbols then we’ll need to do it.

The parser doesn’t know that “NV” is meant to be a constant here rather
than a variable. It doesn’t break anything at the moment because NV is
the only set we ever check membership of, but I assume it’d be a problem
if we also introduced T. This issue will go away if we write raw formula
arrays rather than running strings through the parser.