Make parser generic
Now that we have annotations for nested structure and variables, we can
just turn everything else into symbols — we don’t need to know the
specific structures that the symbols represent.

In turn, this means that we don’t need to distinguish between parsing a
term and formula. These processes are identical now.