Make parser return suspending computation awaiting builder
This is what we will need for rules: we need to rebuild them with a new
scope each time, but the underlying string never changes, so we should
parse it once and then allow an expression to be built (or rebuilt) when
a builder is available later.

For now this behaviour is hidden behind helpers, but next we will want
to reveal it.