Privately instantiate rule premises and conclusion during a match
This makes more sense: if a fresh instance of the rule’s conclusion
succeeds, we use the same scope to instantiate the rule’s premises and
return them along with the post-match state.

For clients this completely hides the details of rule schema
instantiation. Since the premises and conclusion are already private and
only indirectly available through the match interface, it doesn’t matter
that new premises come back every time.